+++ cmd/capi/commands/org_quotas.go
@@ -737,12 +737,12 @@ func newOrgQuotasDeleteCommand() *cobra.Command {
 				return nil, constants.ErrInvalidClientType
 			}
 
-			err := capiClient.OrganizationQuotas().Delete(ctx, guid)
+			job, err := capiClient.OrganizationQuotas().Delete(ctx, guid)
 			if err != nil {
 				return nil, fmt.Errorf("failed to delete organization quota: %w", err)
 			}
 
-			return nil, nil
+			return &job.GUID, nil
 		},
 	}